Vgordievskiy win build (#18)
fixes for windows build

+++ examples/ex_1/main.cpp
@@ -47,7 +47,7 @@ int main()
     double acc = 0.0;
     uint32_t errors = 0;
 
-    uint32_t avarageTime = 0;
+    uint64_t avarageTime = 0;
     uint16_t count = 0;
     std::list<double> items;
     auto bi = std::back_inserter(items);
@@ -92,7 +92,7 @@ int main()
                  | hof::match(print<double>, errorHandler)
                  | hof::filter_if(filter)
                  | hof::match_some(accept)
-                 <<= 0;
+                 <<= 0.0;
 
         auto stop = std::chrono::high_resolution_clock::now();
         auto duration = std::chrono::duration_cast<std::chrono::microseconds>(stop - start);
